if_ntb: Fix build on i386
Sponsored by:	EMC / Isilon Storage Division


git-svn-id: svn+ssh://svn.freebsd.org/base/head@289259 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/ntb/if_ntb/if_ntb.c
@@ -1047,7 +1047,7 @@ ntb_transport_link_work(void *arg)
 	/* send the local info, in the opposite order of the way we read it */
 	for (i = 0; i < num_mw; i++) {
 		rc = ntb_write_remote_spad(ntb, IF_NTB_MW0_SZ_HIGH + (i * 2),
-		    ntb_get_mw_size(ntb, i) >> 32);
+		    (uint64_t)ntb_get_mw_size(ntb, i) >> 32);
 		if (rc != 0)
 			goto out;
